requirements.txt
Django==4.1.2
django-bootstrap5==22.1
Pillow==9.2.0
djangorestframework==3.14.0
